Bug fix/stack over flow and other

+++ nekoyume/Assets/_Scripts/UI/Module/GrindModule.cs
@@ -429,8 +429,8 @@ namespace Nekoyume.UI.Module
             loadingScreen.Show();
             canvasGroup.interactable = false;
             loadingScreen.SetCurrency(
-                (int)_cachedGrindingRewardNCG.MajorUnit,
-                (int)_cachedGrindingRewardCrystal.MajorUnit);
+                (long)_cachedGrindingRewardNCG.MajorUnit,
+                (long)_cachedGrindingRewardCrystal.MajorUnit);
             yield return new WaitForSeconds(.5f);
 
             var quote = L10nManager.Localize("UI_GRIND_NPC_QUOTE");
